+++ ext/pty/depend
@@ -8,11 +8,10 @@ pty.o: $(hdrdir)/ruby/intern.h
 pty.o: $(hdrdir)/ruby/io.h
 pty.o: $(hdrdir)/ruby/missing.h
 pty.o: $(hdrdir)/ruby/oniguruma.h
+pty.o: $(hdrdir)/ruby/process.h
 pty.o: $(hdrdir)/ruby/ruby.h
 pty.o: $(hdrdir)/ruby/st.h
 pty.o: $(hdrdir)/ruby/subst.h
 pty.o: $(hdrdir)/ruby/util.h
-pty.o: $(top_srcdir)/include/ruby.h
-pty.o: $(top_srcdir)/internal.h
 pty.o: pty.c
 # AUTOGENERATED DEPENDENCIES END
